requirements.txt
Flask
cassandra-driver==3.29.1
requests==2.32.3